+++ app/code/ThirdParty/BlogArticle/Ui/Component/Listing/Column/CategoryActions.php
@@ -1,0 +1,49 @@
+<?php
+namespace ThirdParty\BlogArticle\Ui\Component\Listing\Column;
+
+use Magento\Framework\UrlInterface;
+use Magento\Framework\View\Element\UiComponent\ContextInterface;
+use Magento\Framework\View\Element\UiComponentFactory;
+use Magento\Ui\Component\Listing\Columns\Column;
+
+class CategoryActions extends Column
+{
+    private $urlBuilder;
+
+    public function __construct(
+        ContextInterface $context,
+        UiComponentFactory $uiComponentFactory,
+        UrlInterface $urlBuilder,
+        array $components = [],
+        array $data = []
+    ) {
+        $this->urlBuilder = $urlBuilder;
+        parent::__construct($context, $uiComponentFactory, $components, $data);
+    }
+
+    public function prepareDataSource(array $dataSource)
+    {
+        if (!isset($dataSource['data']['items'])) {
+            return $dataSource;
+        }
+        $name = $this->getData('name');
+        foreach ($dataSource['data']['items'] as &$item) {
+            if (!isset($item['category_id'])) {
+                continue;
+            }
+            $item[$name]['edit'] = [
+                'href' => $this->urlBuilder->getUrl('blogarticle/category/edit', ['category_id' => $item['category_id']]),
+                'label' => __('Edit'),
+            ];
+            $item[$name]['delete'] = [
+                'href' => $this->urlBuilder->getUrl('blogarticle/category/delete', ['category_id' => $item['category_id']]),
+                'label' => __('Delete'),
+                'confirm' => [
+                    'title' => __('Delete category'),
+                    'message' => __('Are you sure you want to delete this category?'),
+                ],
+            ];
+        }
+        return $dataSource;
+    }
+}